+++ public/bloblang/spec.go
@@ -86,7 +86,7 @@ type PluginSpec struct {
 	category    string
 	description string
 	impure      bool
-	static      bool
+	isStaticFn  func(params *ParsedParams) bool
 	params      query.Params
 	examples    []pluginExample
 	version     string
@@ -104,6 +104,9 @@ type pluginExample struct {
 func NewPluginSpec() *PluginSpec {
 	return &PluginSpec{
 		params: query.NewParams(),
+		isStaticFn: func(params *ParsedParams) bool {
+			return false
+		},
 	}
 }
 
@@ -209,14 +212,29 @@ func (p *PluginSpec) Impure() *PluginSpec {
 }
 
 // Static marks the plugin as a statically evaluated function or method. This is
-// a guarantee that given the name parameters this plugin will always yield the
+// a guarantee that given the same parameters this plugin will always yield the
 // same value.
 //
 // Marking a function or method as static has the advantage that it can
 // sometimes be optimistically evaluated at mapping parse time when given static
 // arguments.
 func (p *PluginSpec) Static() *PluginSpec {
-	p.static = true
+	p.isStaticFn = func(params *ParsedParams) bool {
+		return true
+	}
+	return p
+}
+
+// StaticWithFunc marks the plugin as a potentially statically evaluated
+// function or method, but only given certain parameters as determined by the
+// provided closure function. This is a guarantee that given the same parameters
+// this plugin will always yield the same value.
+//
+// Marking a function or method as static has the advantage that it can
+// sometimes be optimistically evaluated at mapping parse time when given static
+// arguments.
+func (p *PluginSpec) StaticWithFunc(fn func(params *ParsedParams) bool) *PluginSpec {
+	p.isStaticFn = fn
 	return p
 }
 
@@ -245,6 +263,14 @@ func (p *PluginSpec) EncodeJSON(v []byte) error {
 // instantiation.
 type ParsedParams struct {
 	par *query.ParsedParams
+	e   *Environment
+}
+
+func newParsedParams(p *query.ParsedParams, e *Environment) *ParsedParams {
+	return &ParsedParams{
+		par: p,
+		e:   e,
+	}
 }
 
 // AsSlice returns a slice of raw argument values.
@@ -300,4 +326,10 @@ func (p *ParsedParams) GetBool(name string) (bool, error) {
 // defined, otherwise nil.
 func (p *ParsedParams) GetOptionalBool(name string) (*bool, error) {
 	return p.par.FieldOptionalBool(name)
+}
+
+// ImportFile attempts to read a file via the underlying environment importer.
+// Relative paths will be resolved from the path of the file being imported.
+func (p *ParsedParams) ImportFile(name string) ([]byte, error) {
+	return p.e.env.ImportFile(name)
 }