+++ internal/impl/pure/processor_log.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/benthosdev/benthos/v4/internal/bloblang/field"
 	"github.com/benthosdev/benthos/v4/internal/bloblang/mapping"
-	"github.com/benthosdev/benthos/v4/internal/bloblang/query"
 	"github.com/benthosdev/benthos/v4/internal/bundle"
 	"github.com/benthosdev/benthos/v4/internal/component/processor"
 	"github.com/benthosdev/benthos/v4/internal/docs"
@@ -136,20 +135,15 @@ func (l *logProcessor) ProcessBatch(ctx *processor.BatchProcContext, msg message
 	_ = msg.Iter(func(i int, _ *message.Part) error {
 		targetLog := l.logger
 		if l.fieldsMapping != nil {
-			v, err := l.fieldsMapping.Exec(query.FunctionContext{
-				Maps:     map[string]query.Function{},
-				Vars:     map[string]any{},
-				Index:    i,
-				MsgBatch: msg,
-			}.WithValueFunc(func() *any {
-				jObj, err := msg.Get(i).AsStructured()
-				if err != nil {
-					return nil
-				}
-				return &jObj
-			}))
+			fieldsMsg, err := l.fieldsMapping.MapPart(i, msg)
 			if err != nil {
 				l.logger.Errorf("Failed to execute fields mapping: %v", err)
+				return nil
+			}
+
+			v, err := fieldsMsg.AsStructured()
+			if err != nil {
+				l.logger.Errorf("Failed to extract fields object: %v", err)
 				return nil
 			}