+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - Fixed a potential panic when closing a `pusher` output that was never initialised.
 - The `sftp` output now reconnects upon being disconnected by the Azure idle timeout.
 - The `switch` output now produces error logs when messages do not pass at least one case with `strict_mode` enabled, previously these rejected messages were potentially re-processed in a loop without any logs depending on the config. An inaccuracy to the documentation has also been fixed in order to clarify behaviour when strict mode is not enabled.
+- The `log` processor `fields_mapping` field should no longer reject metadata queries using `@` syntax.
 
 ## 4.20.0 - 2023-08-22